Revert datCalcGetHp widening: broke datGetHp MATCH in g_data.c with zero benefit to its target

+++ include/Main/Battle/Data/datCalc.h
@@ -19,7 +19,7 @@ typedef struct DatUnit DatUnit;
 u32 datCalcRand(u32 max);
 
 u8 datCalcGetLevel(DatUnit* unit);
-s32 datCalcGetHp(DatUnit* unit);
+u16 datCalcGetHp(DatUnit* unit);
 u16 datCalcGetSp(DatUnit* unit);
 void datCalcSetHp(DatUnit* unit, u16 hp);
 void datCalcSetSp(DatUnit* unit, u16 sp);
